requirements.txt
requests
pandas
pyyaml
pytest
cachetools